Implement a map version of thrifty and convert PortMap
Summary:
Adding ThriftyNodeMapT which is a map version of thrifty and has a lot of to/from thrift code built in. With this it should be easier to convert maps of thrifty nodes to thrifty. As an example I added this to PortMap, so now SwitchState+PortMap can be converted to thrift.
Some unknowns which I still have to figure out are the extraFields in NodeMapT and having more complex keys for the map (right now I'm just using static_cast)

Reviewed By: aeckert

Differential Revision: D30946359

fbshipit-source-id: 4ff46a6d1f5d2eafef1c57e3a01213c2ca238928