Implement a map version of thrifty and convert PortMap
Summary:
Resubmitting D32029247 (https://github.com/facebook/fboss/commit/2dbc2ef0159e6bfbc16435f9680f570a000783ab) (reverted in D32029247 (https://github.com/facebook/fboss/commit/2dbc2ef0159e6bfbc16435f9680f570a000783ab)). This caused a crash at warmboot because, while to/from follyDynamic was working fine, the dynamic that generated was not json serializable. The dynamic I was generating was in the form {"sw": "ports": { 1: {<port_data>}}}. Json object keys need to be strings so this is invalid due to the 1 being an int. Changed the key to a string.

Original:
Adding ThriftyNodeMapT which is a map version of thrifty and has a lot of to/from thrift code built in. With this it should be easier to convert maps of thrifty nodes to thrifty. As an example I added this to PortMap, so now SwitchState+PortMap can be converted to thrift.
Some unknowns which I still have to figure out are the extraFields in NodeMapT and having more complex keys for the map (right now I'm just using static_cast)

Differential Revision: D32035313

fbshipit-source-id: fbf287220c8785881b206bd412fbd69ce495fcbf