Support updateStateBlocking
Summary:
Because current TransceiverManager::updateState() is only in charge of creating TransceiverStateMachineUpdate and sending
it to the pendingUpdates_ queue **without** waiting for the result.
This is fine since our current queue can make sure the order of the upcoming state.
But we don't want to see the refreshTransceivers() function calling triggerProgrammingEvents()
to program the same xphy ports for the same transceiver multiple times even though the last
update is still in the queue or running(because port programming can be slow).

Therefore, similar to wedge_agent, I introduce the `BlockingTransceiverStateMachineUpdate` to support updateStateBlocking()

Reviewed By: birdsoup

Differential Revision: D31866633

fbshipit-source-id: ee2c0c351e3290c63a8e9200bc84522c17278b77