Added module media interface to transceiverInfo
Summary:
facebook
module interface id is the main media type that the optic was designed for. The current mediaInterfaceId that we export in transceiverInfo is the operational media interface.
For example, if a 200G-FR4 is configured to run at 100G, the current transceiverInfo will show the mediaInterfaceId as 100G-CWDM4. While that is correct, for NGT we also need to know what the actual mediaInterface is, in this example - 200G-FR4.

For sff/sff8472, this is straightforward. The modules don't change application with speed.
For cmis, in order to get the main module interface id, I am currently looking for the highest supported speed and the corresponding application. Assumption (verified with hw team also) - A single optic will not advertise two applications at the same speed so we should be fine with looking at the max application

The moduleMediaInterfaceId will be required in NGT for picking correct thresholds for optics

This change adds the moduleMediaInterfaceId to transceiverInfo

Differential Revision: D32009292

fbshipit-source-id: 984d120c5fc27337dc08f8e8d479d1850ed3483a