Provide execCommand and execCommandUnchecked abstractions
Summary:
In the common case, we want to check return status
of command and fail hard (via exception) in case the command
fails. Make this the behavior of execCommand. This saves us
from repeating status check code. Provide another api -execCommandUnchecked
where we return exit status and leave it to the client to
check status

Reviewed By: binhuang00

Differential Revision: D32153295

fbshipit-source-id: 846cfffa32cde2830a84bb9b5d87dcf2478e73e5