Support updateTransceiverPortStatus()
Summary:
Right now, we rely on wedge_agent to call syncPorts() so that qsfp_service can know what's the status of all the ports of the transceivers. And this function call will only be called in the link event. If qsfp_service misses the last call during restarts, it will
never have the accurate port status until the new link event occurs.

As we're moving the whole port programming to qsfp_service to control when's the best time to program iphy/xphy/tcvr and remediate
tcvr. I think it will be a better idea to let qsfp_service to ask wedge_agent for the real time port status.

Therefore, after we program transceiver in the state machine, we will call wedge_agent's getPortStatus() to get the latest status
of all ports. And then it will only update the state machine for the stabled ports, which are not recently programmed.

Reviewed By: birdsoup

Differential Revision: D31999805

fbshipit-source-id: 087cb5db68b25037b4b9ac62145d15d62c109db5