Use nexthop members as well as weights to identify NextHopGroupTraits
Summary:
The currennt implementation uses set of nexthops as adapterHostKey for
SaiNextHopGroupTraits.

As a result, two routes with identical nexthops but different weights end up
with single entry in the SaiStore. This in turn causes these two routes to
point to same ECMP group instead of different groups.

Fix it by using set of pair of nexthop & weight as key.

Reviewed By: msomasundaran

Differential Revision: D32229488

fbshipit-source-id: 2476bc5392597624167b3ab2b4cb9214652db983