Check if smfCode exists in the map before accessing it
Summary:
mediaInterfaceToPortSpeedMapping is an unordered_map and if a smfCode doesn't exist, it adds a default node for that key. This could result in race conditions with one module's thread inserting a node in this map, and the other module's thread reading from this map.
It's better anyways to confirm first if the key exists and then access it.

Reviewed By: joseph5wu

Differential Revision: D32327438

fbshipit-source-id: 099e6566b94cb0b63f140d63cbe294e3f00eb428