Add NeighborUpdaterAgentTestImpl
Summary: Added a stub version of NeighborUpdaterImpl to disable the NeighborUpdaterImpl which takes care of creating and maintaining state machine for NDP/ARP entries. Idea is to be able to use the stubbed version in agent tests where we don't want these entries to be created as they cause traffic on CPU queues and interfere with agent test runs.

Reviewed By: aeckert

Differential Revision: D32286209

fbshipit-source-id: 707ce07cb350ad14fdd1480c2cb1b4fd8f84ee24