enable afe trim to adaptive during oper up
Summary:
As titled, set afe trim to ADAPTIVE on tajo platform when oper up is received on a port from the SDK.

This is to mitigate S249471.

Apply the setting only when:

- port media type is copper
- port speed is 25G
- port oper up from the SDK

When a port is unstable due to afe mode being fixed, port flaps. Use the oper up event from the SDK to configure the port to adaptive.

Next steps:

- Push this change to the fleet through warmboot before scheduling coldboots on 400C.
- Once every device in the fleet is coldbooted, this routine can be removed.
- Note that the function ensures to set AFE trim to adaptive only when its not set before

Differential Revision: D32411333

fbshipit-source-id: 3fd99228459a3b5911a18a8f08905983d1ee1d48