Add multiple LAG ports to test ECMP
Summary:
Previously we had 2 ports in multinode tests and both ports were part of a single LAG. Now we have 2 additional ports using which i am creating another LAG. With 2 lags we get a combination of ECMP + LAG which allows us to check for ECMP shrink during warm boot.

Summary of changes
1. New api in config factory to create multiple L3 interfaces with a given number of ports per L3 interface
2. Add a second LAG between the switches
3. Check for ECMP shrink after warmboot.

Note:
With 2 LAGs, i see few initial packets getting dropped as in discards when starting traffic though default route and ARP/NDP entries are present. The drops happen for the initial few packets and do not happen after that. I am following up on this with broadcom using a CSP case. Till then the in discard verification is limited to post warmboot phase

Differential Revision: D32736946

fbshipit-source-id: 27e07652662ccae3b73760eb9458da640624d9a9