Adjust bringup time for remote side
Summary:
For testing LACP, we wait for switch state changes. Till now the state check was needed only on UUT. The remote side is passive and does not do verification.

For disabling TTL decrement on nexthop, the remote side has to wait till switch state is setup. Netcastle spawns the binary on remote side first and then starts the binary on UUT. So remote switch has an additional wait time which is equal to the boot time on UUT.

The tests work with current wait time of 60 seconds on remote switch. But it is better to account for additional bringup time on UUT. Hence adding a separate retry count for remote side

Differential Revision: D32838215

fbshipit-source-id: d115c14fa8b8dc02d36f76d5500495c25806a0a8