switch NeighborCacheEntry::probesLeft_ to a 32-bit int
Summary:
We unexpectedly were seeing neighbor cache entries moving to
EXPIRED under a minute after the neighbor goes away, even though
maxNeighborProbes was set to 300.

The reason is that internally we still had this variable as a uint8_t,
meaning the max value is 255. When we set it to 300, we end up
overflowing and effectively storing 44 instead.

This diff changes the value to be a 32 bit integer to avoid this
issue and allow setting values >255.

Reviewed By: jasmeetbagga

Differential Revision: D32908319

fbshipit-source-id: 81e5c7c5419e9664200a92c4df682e70b527ca60