a method to return list of ASICs
Summary:
as titled.

A utility method in HW test to return list of all ASICs and also provide a list of ASICs not same as the one on which HW Test is running.

This could be useful to emulate behavior of two interlinked switches of different ASICs.

Reorganized enumerated list with macro so list of ASICs can be added/extended at same place.

Reviewed By: joseph5wu

Differential Revision: D32776458

fbshipit-source-id: 9297e13cc546711fc69bfb290715d5980e935441