avoid adding acl entry with duplicate priority
Summary:
methods to create acl entry for trapping packets were being created with priority 1. this assumes that no other acl entry with that priority exists. but in default config, copp acls may be present. this leads to exception while adding an acl with priority 1.

to prevent this, check how many acls exist and add with a priority not used by other acl entries.

Reviewed By: shri-khare

Differential Revision: D33009369

fbshipit-source-id: 55c68266d4f1c3bfd8e25e7a56c20421116e75b2