Move updateTransceiverPortStatus() before any programming event
Summary:
In this diff, I decided to move the fetching port status from wedge_agent the first step of the whole refresh transceiver state machine logic.
The main reason is that some of the following step, like removing a transceiver or remediating a transceiver
might depend on whether there's still some port up.
And if there's port up, we shouldn't remediate the transceiver.

Besides, without moving this to the first step, it won't fix the issue that we disable ports on wedge_agent and then enable these ports later.
In this way, we won't trigger an agent config change.
So if the first time we program internal phy port with disabled state, we won't cache any disabled ports in the `tcvrToPortInfo_` map.

Therefore this new updateTransceiverPortStatus() will have a bit more complicated logic to decide whether there's a port change and then
to decide whether it's OK to just trigger a port status change event or a reset to not present/discovered event, which will trigger a new
port programming if necessary.

Reviewed By: birdsoup

Differential Revision: D32518826

fbshipit-source-id: 15d098b57cbd7245986cd675bd0e683aa2529385