Fix the blocking update during exiting
Summary:
Current `updateStateBlocking()` logic has a bug that if the system is in existing mode, the update will just return without actually adding to the update queue:
https://www.internalfb.com/code/fbsource/[33226e83e45eaf0bc10bf2c71c06f9910d57c79a]/fbcode/fboss/qsfp_service/TransceiverManager.cpp?lines=186-192

Therefore, if we still wait for the BlockingTransceiverStateMachineUpdateResult, it will never flag to success or error.

To fix that, I change the above function to return true or false to represent whether the update actually added to the queue.
We will only wait the result when the update is added to the queue.

Besides, I also replaced the regular update function:
```
void updateState(TransceiverID id, TransceiverStateMachineEvent event);
```
to
```
  void updateStateBlocking(
      TransceiverID id,
      TransceiverStateMachineEvent event);
```

Since all operations in one transceiver should be sequential and needs proper order, we should just use the blocking update to make
sure the last update is done before moving on to the next update.

Differential Revision: D32572721

fbshipit-source-id: 9aed57197d5c138a9a94563e332359d10026ce63