Th4 support for route counters
Summary:
Adding route counter support for TH4.

On tomahawk3 devices, route counters are allocated dynamically and attached to routes. The initial plan was to follow the same model on TH4 with flexCounters. However i ran into issues when allocating multiple flex counter actions and broadcom identified this as hardware limitation. As a work around they suggested creating a single flex counter action for all routes. The flex counter action has multiple offsets which can be used as individual counters. This creates multiple limitations on TH4.
 * We have to preallocate max counters
 * If max counters change, the following sequence is needed.

      1) detach all routes pointing to existing counter action
       2) deallocate counter action
       3) create new counter action with max counter
       4) reattach all counters

This increases the complexity in agent especially if there are failures during warmboot. The number of counter IDs is defined in switch configuration and is enforced in software. The hard limit for counters currently is 255 based on hardware capacity of Cisco chips. We will allocate 1k counters on TH4 always and enforce the max counterID defined in switch config by checks in software. This allows us to handle maxCounterID changes for values up to 1k without having to reallocate the block.

 * V6 and V4 prefixes cannot share counters. The DC controller use case include only v6 routes. So i am excluding v4 routes to save counter space.

Summary of changes:
- Converted BcmRouteCounterID from an int to a struct to accommodate index and offset for TH4
- Moved counter and counter table common functionality to base objects and created child objects for flex and th3 counters and table management.

Differential Revision: D30078677

fbshipit-source-id: afc7202035095fa42945119cfb925c63cf1337e6