Fill in any 'null' tx settings with defaults
Summary: PortProfile tests were failing on fuji due to the profiles not matching. On BCM xphys we can just leave the expected TX settings as null to not program them, but when we read them back from the HW we'll read the default values. Because of this, we need to fill the expected config in with the default TX settings in place of any 'null' values before comparing it to the actual config.

Reviewed By: joseph5wu

Differential Revision: D32301346

fbshipit-source-id: 91f96fe91bdea4797b11e7552ad7f9e54c6c2ddc