Don't set COS MAP for CPU port for BRCM
Summary:
DSCP to Queue and Queue to TC mapping is a per port config.  When FBOSS
sets it on a particular port, the expectation is that this configuration
will take effect for packets ingress on that port.

However, when FBOSS applies this configuration on the CPU port, a bug in
BRCM-SAI means that the configuration is applied for packets *egress* to CPU
port.

We are discussing with BRCM on how to fix this issue, however, the fix looks
involved (requires several FBOSS side changes) + it may only be available on
BRCM-SAI 6.0 (we need this fix for 4.2, and it appears like a non-trivial
effort to backport 6.0 fix to 4.2, those discussions are underway). This would
be the longer term solution.

Ih the meantime, skip setting this configuration on CPU port:
 - By default, BRCM SAI maps TC 0 through 9 to Queue 0 through 9 respectively.
 - OpenR ACL has action to set TC to 9, so in conjunction with the default TC
   to Queue mapping, it could steer OpenR packets to the right queue (high
   priority queue 9).
 - Packets ingress from CPU port continue to follow Olympic model since
   CPU port shares the cos map configuration on other front panel port
   (BRCM ASIC behavior).

Note: BRCM-SAI implements SAI HostIfTraps using ACLs or RxReasons code.
L3_MTU_ERROR is the only FBOSS HostIfTrap implemented using RxReasons. That is
broken today, and continues to be broken even with the workaround in this fix,
the longer term fix woudl address this.

Differential Revision: D33050811

fbshipit-source-id: d55ee48d20f7b7c86a434b5f6a41cda6336898fa