a method to setup a warmboot handles
Summary:
consider a warm boot case for acl table and acl table schema is changed.

as per tajo, we need to perform following
1) take back up of all old acl entries
2) remove acl entries
3) remove acl table
4) create acl table with new schema
5) create acl entries in new table

for a warm boot, we need a mechanism to remove a warm boot entry and then replace it (recreate a new one to replace it) in warm boot handles. this must be done before an entry is already claimed,

providing a method to setup warm boot handle.

Reviewed By: jasmeetbagga

Differential Revision: D33179169

fbshipit-source-id: 70bc384f303b1bcb628b0953dabaa5933c4a4b6a