do not swap afe trim and dsp mode if its a warmboot
Summary:
As titled, do not swap afe trim and dsp mode if its a warmboot. This is a temporary solution that we will be rolling out till AFE trim fix is coldbooted and will be removed after rolling out AFE TRIM (S249471)

 The DspMode and AfeTrim values provided by vendor are swapped (incorrect) for 25G serdes. With this diff stack (D31939792 (https://github.com/facebook/fboss/commit/9f82003b2e281cac3f5dbde23515ede29484ea39)), it is rectified but this cannot be rolled out during warmboot since this will cause port flaps.

Hence, swap the values back if its a warmboot. As we are scheduling coldboots for AFE TRIM SEV S249471, utilize this to propagate the correct values.

This will help us carry on the warmboot upgrades on 400C devices while we wait on coldboots to happen over a period of 3 months.

     * --------------------------------------------------------------------
     *               | PIN DSP | PIN AFE | STORE DSP | STORE AFE | ACTION |
     * --------------------------------------------------------------------
     * WB before CB  |    7    |    4    |     4     |     7     |  SWAP  |
     * --------------------------------------------------------------------
     *       CB      |    7    |    4    |     -     |     -     |  SKIP  |
     * --------------------------------------------------------------------
     * WB after CB   |    7    |    4    |     7     |     4     |  SKIP  |
     * --------------------------------------------------------------------

Differential Revision: D33243678

fbshipit-source-id: 86c315cfb154653d148f0ab52331ce99aa316839