Sai Replayer extension attribute support
Summary:
As titled. Adding extension attributes to the attribute map as well, if extension attribute is present. The check needs to be done at runtime, since Sai Replayer does not know beforehand if the attribute exists for different sai implementation.

`handleExtensionAttributes()` will take care the logic of adding extension attributes to the map.

Differential Revision: D33250899

fbshipit-source-id: 599921ae4d04657ed1b2cd03274ef898553178b8