Make sure update ACTIVE state when port status change
Summary:
Based on the comments w/ birdsoup, we agreed that we should keep current agent -> qsfp about port status change logic via syncPorts(),
so that we won't accidentally think a flapping port is a hard down port due to the 5 seconds routine status check from the state machine.

Therefore, changing this diff to update the state machine state and the cached port status when syncPorts() is called.

Reviewed By: birdsoup

Differential Revision: D32904076

fbshipit-source-id: b5136aae0619cf9f8463d7267ffcc74f5b7856ad