Publish xphy and tcvr snapshots when there's a port changed
Summary:
Several refactor inside WedgeManager:
- Refactor the current `getTransceiverID()` in WedgeManager to just use the cached `portToSwPortInfo_` map
which is generated in the TransceiverManager constructor by using PlatformMapping.
Therefore, we don't have to look through the platform mapping again and again using `utility::getTransceiverId()`

- Also refactor the current WedgeManager::publishTransceiverSnapshots() to TransceiverManager::publishLinkSnapshots(), which we will
also try to publish the xphy snapshots for a port if qsfp-service is using the new port programming logic.

- Finally add publishLinkSnapshots() in the current two places which will change port status in the new state machine:
  - TransceiverManager::updateTransceiverPortStatus()
  - TransceiverManager::updateTransceiverActiveState()

Reviewed By: birdsoup

Differential Revision: D33183099

fbshipit-source-id: bc6f71fe626b259e354f9f755b408279fcba4c31