Wait until all modules READY in cmisStateChanged test
Summary:
It appears that sometimes the modules dont go into the READY state very quickly after programming.

This causes the next refresh() cycle to still have state changes, and will cause us to fail the test when we check that the stateChange flag isn't asserted.

This diff changes the behavior to instead refresh() until all modules are in the READY state before we assert that the state doesn't change on future refreshes.

Reviewed By: joseph5wu

Differential Revision: D33290052

fbshipit-source-id: db9dbc9a184a624a54ac3d6659138a1f14d56837