Increase the wait time for WRED tests to fix the flakiness
Summary:
In the current verifyWred test, the loop seems to run for ~1s for TH3
and hence WRED drops are not seen sometimes in this period, failing the test.
To reduce the flakiness, it would be good to explicitly specify the number of
iterations and the wait needed bw iterations to allow say a max of 4 seconds.

Reviewed By: jasmeetbagga

Differential Revision: D33445416

fbshipit-source-id: 2bfe16b1314f697c0bbb4b2d857739d79e063295