Fix prbs test warmboot issue when using --new-port-programming
Summary:
There's a bug in current PhyManager::programOnePort(), which always resets PortToExternalPhyPortStats mapping
no matter whether there's any change on the profile/speed.
This will be an issue for warmboot, because it will lose the prbs stats collection setup.

To fix that, we will return true or false for setPortToPortCacheInfoLocked() to decide whether there's a change on the cache of xphy ports.
And then use that to decide whether we need to reset the stats.

Differential Revision: D33479454

fbshipit-source-id: 005e5da2a12f23fd6ff15d3cbd4f3694e161d710