Fix link xphy port stats test when using --new-port-programming
Summary:
With the new port programming enabled, we will move xphy programming into qsfp_service,
therefore, we can't just use dcData() from link_test, which is actually running as wedge_agent,
to fetch the xphy related stats.

To support the new port programming, we need to ask qsfp_service to get xphy related counters.

Besides, current xphySanity check is using a try-catch to have a better log info if the test failed on the checkWithRetry(),
I decided to add a new input: conditionFailedLog for this common function and simplify the current test logic

Differential Revision: D33489266

fbshipit-source-id: 09fcd42229f65befcb47a666d1e4d11bf7f29009