Start building basic fsdb integration tests
Summary:
Simple test that creates a subscription for wedge_agent stats with
FSDB. This simple test revealed a few limitation
- ServiceRouter doesn't work well on our lab boxes. It croaks for want of all
  the fbwhoami fields required by ServiceRouter -
  https://fburl.com/code/eu1smd87.
- We are using TLS even with --thrift_ssl_policy=permitted passed in
- SSL should always be bypassed for localhost connections. It isn't for FSDB.
This is a must for all critical services.

For the last 2 I had to explicitly disable thrift security in connection setup
in FSDB Client (viz. connConfigs["thrift_security"] = "disabled"). We will need
to fix all these for prod. I will block landing the diff till
a) We resolve these
b) We get netcastle to start running FSDB in link tests

Differential Revision: D33246651

fbshipit-source-id: aa721a4ecf2652dd74ff5fd142523a1b7a43182d