Access parity error counter via service data and not TL stat var
Summary:
- Trigger aggregation of TL stats since we don't run a stats
publish thread in HwTests

- Access counter value from service data. We should not be
accessing thread local vars and asserting on their value.
The TL var thus accessed would be local to the thread
and thus only represent the value on that thread.

As the next diff shows, we are actually using TL stats improperly.
These were contained in a non thread local HwSwitchStats. This
created a single copy of such TL stats (created in the thread
HwSwitchStats was created in), which then all threads would access
via the HwSwitchStats pointer. This is not correct usage. Next
diff fixes that.

Differential Revision: D33811224

fbshipit-source-id: 159430715aaefa88c89152f08a9d5a413cb6e3cf