Unify handling of SwitchStats, HwSwitchStats and fix access race
Summary:
HwSwitchStats is accessed from multiple threads without
any synchronization. This happens to be ok since we are only
accessing TL stats via HwSwitchStats object. But if we ever
accessed anything else, it would be racy. We have a similar
pattern in SwSwitch::SwitchStats, so unify storage for
both to thread local

Differential Revision: D33795601

fbshipit-source-id: cdf787f5d54420bc9e13bc10b2799cfc8c1c0a76