Add API to read the capture routines for count of packets in buffer.
Summary:
(1) This utility is needed in test, so we do not end up stopping the capture file too soon (before enough packets are captured). We want to wait for some deterministic time and we use this count to determine progress of the pcap buffer.
(2) readPcapFile has been added an option to skip exception. Apparently it seems to fail in read of last packet intermittently, but it doesn't effect captured pkt .

Differential Revision: D33809275

fbshipit-source-id: 98d17b6542e85d1401fc9a9f9af1387648c3b6f7