Get an unused priority for ACL entry in test
Summary:
When a new entry is being added with a priority, need to make sure an
unused priority is used to add the entry to avoid add time failure due to
duplicate priority. Adding a function to find the next free unused priority to
use for next entry being added.

The specific case thats failing is below:
- Test : HwCoppQosTest.HighVsLowerPriorityCpuQueueTrafficPrioritization
- ACL entries created with priorities 2-5, number of entries = 4, P471587031
- New trap entry addition fails trying to add priority 5 (# of entries +1), P471740110
- Full logs [here](https://www.internalfb.com/intern/everpaste/?color=1&handle=GDc8bgp2IR5rx08DAHQ2n_jXn_cObr0LAAAz)

Reviewed By: phshaikh

Differential Revision: D33115620

fbshipit-source-id: 0814ae55d34447c0417623c1777fe5456cb7d701