Implement basic filter functionality
Summary:
Until we have a the full robust --filter functionality we have planned, I think its good to have at least some basic manual filter feature in case command implementations need it.
With this commands can define a FILTERS array and specify which keys they will accept. The framework will check to make sure the correct keys are passed and in the right format but does not do any validating on the filter values for now.
Implemented state filter for `show port` as a toy example

Reviewed By: hegara

Differential Revision: D32851107

fbshipit-source-id: 6e65824706203929d59a519fe1633d793ce32a85