Introduce getConfigAppliedInfo()
Summary:
As mentioned in T111079963, we need to inform qsfp_service when there's a coldboot in wedge_agent,
so that when wedge_agent finished reset switch asic, we need to re-initiate xphy and transceiver data path.
In such way, we can make sure the high speed optics can get a clean signal from iphy/xphy to avoid stuck in the bad state.

This diff will introduce a new thrift api getConfigAppliedInfo() to replace current getLastConfigAppliedInMs() which we can't tell whether
the last config applied is during coldboot.

Reviewed By: jasmeetbagga

Differential Revision: D33900353

fbshipit-source-id: 5b91cfb79eee30c3bba5ae4fc3349944f484b1d2