Add mpls in Invariant test
Summary:
The intent of this change is to get prod to trunk and viceversa coverage for mpls rib feature enablement. mpls currently works without rib and it will transitioned to use rib soon. So we need to make sure that warmboot in either direction works fine.

Adding mpls test to invariants. This will run on prod version with rib disabled and while enabling rib will exercise the same test with rib.

Differential Revision: D33835278

fbshipit-source-id: c9f38df51e706f6c5c183fafeb7744775600f331