Add a convertor for mpls rib/no rib conversion
Summary:
Even though we use the RIB route struct in label FIB currently, there is a difference in how the struct is populated when RIB is enabled/disabled.

- without RIB, thrift layer fills in the interface id and saves client entries with interface id.
- with RIB, the client entries are stored as it was received from the producer. This means that only v6 linklocal and interface routes have interface id. Rib asserts it here - https://fburl.com/code/9fusxura

Adding a utility to convert from no rib to rib warmboot by reseting the interface for non linklocal entries. When warmbooting from rib to no rib case, it is not possible to fill in interface ids. But this should be fine since we have only one client entry for mpls routes and any update to that route should come with interface id from thrift layer.

Reviewed By: jasmeetbagga

Differential Revision: D33835279

fbshipit-source-id: 6c5de083d22ea90891b93b5e7112da478a2123bd