Fix the data race issue in FbFpgaPimQsfpController
Summary:
Because all the qsfp with the same controller is sharing the same offset,
and use different bits to read the value, it will be racy for different
qsfps try to read the same register offset value and then write at the same
time. Use mutex to protect the race condition for each register offset.

And then use lock_guard() to guard each register offset before read/write the value.

Reviewed By: birdsoup

Differential Revision: D34018328

fbshipit-source-id: 5464da92fa1f0d58223298b32b9fec157be94bfc