Move common LED mode code to ledUtils
Summary:
We need to enable the LED mode both in sai and in native wedge100 code. Right now we're only doing it in the native code, so I'm moving it to a common function so we can use the same logic in our SAI implementation

This should fix issues where the LEDs are stuck in "party mode" after a power cycle.

Reviewed By: phshaikh

Differential Revision: D33986588

fbshipit-source-id: 863dd4d7d1ed3691048016a13bbdc06b5fc70bff