Correct eye calculation for BCM, disable eye calculation for Credo
Summary:
For BCM, the eye should be the sum, not the difference.

For Credo, it sounds like the eye margin we're getting is the height, not the width. But with the test I added, I found that even the old code before NGT can give negative height. The eye margin code seems to be doing a diff just like I did for BCM. We may need to change it into sum also but it's not clear if there're existing consumers we may affect. We'll just comment it out for now while we figure this out so NGT don't freak out about the negative values.

Differential Revision: D33990588

fbshipit-source-id: f832f63c01196f4db39337ad3d003c3d04b77a51