Add API to look for a min instead of 0 watermark
Summary:
Device watermark for TAJO will always be non zero as there
is always some system buffers being used, even with no traffic. Hence
we need to modify the assumption that without traffic, device
watermark will be zero. The expectation for TAJO is to see device
watermark in the range of 38KB without traffic. Now, as this is the
new zero threshold, we will need to modify device watermark tests
to use > 38KB by causing some kind of congestion to ensure buffer
usage is going higher. This change adds a new API which instead of
looking for zero, look for if the usage is <= the specific zero
threshold, 38K for TAJO, 0 for rest of the platforms

Reviewed By: jasmeetbagga

Differential Revision: D34030116

fbshipit-source-id: 6da73ac4e7a5eb862aa33ba6777e3bd64e25a8ee