Create thrifty versions of [Arp|Ndp]ResponseEntry
Summary:
NeighborResponseTable is a little bit more of a problem to convert to Thrifty than previous map types. The complication comes from the fact that its not a NodeMap type but actually normal NodeBase that encodes a map. This also means that we're not really getting any shared_ptr comparison per entry but rather on the table level since the table is holding entries by value.

I want to convert this to a ThriftyNodeMap to keep consistent with the rest of our state, get shared_ptr comparison for deltas down the line and so we can inherit the toThrift/fromThrift. This poses a little bit of challenge for the unit test picture. In past maps I used validateThriftyMigration which serialized/deserializes back and forth using new/legacy serialzers.

Once I convert to thrifty I can't reuse the legacy serializer to test so to get full conversion coverage I've opted to first create a mirror of the tables in thrifty and unit test the conversion, then in the next diff convert the actual table and get rid of the temp code

Reviewed By: jasmeetbagga

Differential Revision: D33903904

fbshipit-source-id: 72803f7c4f1df733fd911691891aa00d6283e215