Add getSensorsData to sensor service to retrieve multiple sensors
Summary: The currently provided APIs for the sensor service provide information about either a single sensor, or all sensors. This changeset adds an additional method to obtain information about a provided set of sensors. It then uses this in the sensor Thrift service to provide a more efficient mechanism for fetching details about multiple (but not all) sensors at once.

Reviewed By: binhuang00

Differential Revision: D34078325

fbshipit-source-id: a18af96584c189d36c804a64191258d21d0875db