Introduce HwStateMachineTest::CheckAgentConfigChanged
Summary:
This diff introduced a new HwStateMachineTest: CheckAgentConfigChanged.
This test will mimic the logic of wedge_agent warmboot and coldboot, which both will cause a config reloaded, and the applied time will be updated
accordingly.

The tests will verify the order of the TransceiverStateMachineState after we set the overrideAgentConfigAppliedInfoForTesting_,
since we don't have wedge_agent running for qsfp_test to call the thrift apis.

Reviewed By: birdsoup

Differential Revision: D33991291

fbshipit-source-id: 0c23434290d80ebb9d90558d39eb9fd94561b48e