Add packet observer mechanism for pcap and agent tests.
Summary:
We already use observer mechanism for the (1) stateUpdates in SwSwitch.cpp (2) for the packet Rx, ink callbacks etc in HwEnsemble.
This observer mechanism is same but more generic to be used anywhere in fboss . For now it handles only the Packet Rx callback but can be extended for anything else as well.

- Decoupled it from SwSwitch so that PcapManager can also used it without swSwitch dependency.
- Decoupled from HwEnsemble so that Agent Tests can use it. This will be leveraged so that these tests can snoop the pkts of interest.

Reviewed By: jasmeetbagga

Differential Revision: D33948050

fbshipit-source-id: b14e4302af1c0d2b38c539dcdfa6efd42c153c4c