Allow for cmd line flag override before switch creation in tests
Summary:
Currently we had a api AgentTest::setupFlags that tests
could use to override cmd line flags selectively. This is most
useful when you want to test a particular behavior, but are not
ready to turn it on for all tests (e.g. fsdb publishing).
The problem with this api was that it was called post switch
creation. So the flags were not in effect during switch creation.
To work around this, some tests use SetUp to set their command line
flags. This was problematic on 2 counts
- Created 2 mechanisms to do the same thing
- It didn't work. Since AgentInitializer::createSwitch would parse
cmd line flags from config and set all else to their defaults.

So create a mechanism that allows early override of flags

Reviewed By: msomasundaran

Differential Revision: D34223105

fbshipit-source-id: 131bcdbb17f6330107518769bc6a4f4189740912